Finish implementing the project's UI/UX
This change refactors public/index.php registering the three required
routes, and adding template support using Twig as the template engine.
It also adds the templates and supporting images so that each of the
routes can be viewed, even if they don't do anything yet. The intention,
of this commit, is to allow the application to be viewed, so that you
can get an idea of what it will eventually do, even if it can't
currently do that.