Add QA tooling support
This change adds the standard QA tooling of PHPUnit, PHP CodeSniffer,
and PHPStan that I used with most of my projects (if not all of them).
It updates the scripts configuration and adding scripts for each one of
them, adds the required configuration files. Finally, it adds a few
other configuration options to composer.json, so that the configuration
is more complete and easily maintainable over time, and fixes the
autoload configuration.