Add and enable mezzio/mezzio-flash as a project dependency
This change adds mezzio/mezzio-flash as a project dependency, registers
the required services, by adding its ConfigProvider and the ones from
both mezzio-session and mezzio-session-ext and adding both
SessionMiddleware and FlashMessageMiddleware to the application's pipe.
Thanks to this configuration, flash message support will be available to
all routes, without route-specific configuration.

Flash message support is required to rapidly add the ability to work
with flash messages in the application. This is required to let the user
know when things go wrong, mainly, as well as when things have worked
successfully, such as file uploads.